# app/static/swagger.json
{
    "swagger": "2.0",
    "info": {
      "title": "Blog API",
      "description": "API for managing users and blog posts.",
      "version": "1.0.0"
    },
    "host": "localhost:5000",
    "schemes": ["http"],
    "basePath": "/",
    "produces": ["application/json"],
    "paths": {
      "/auth/register": {
        "post": {
          "tags": ["Auth"],
          "summary": "Register a new user",
          "description": "This route registers a new user with a username, email, and password.",
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "description": "User data",
              "required": true,
              "schema": {
                "$ref": "#/definitions/UserRegister"
              }
            }
          ],
          "responses": {
            "201": {
              "description": "User registered successfully",
              "schema": {
                "$ref": "#/definitions/UserResponse"
              }
            },
            "400": {
              "description": "Invalid input data"
            },
            "409": {
              "description": "Email already in use"
            }
          }
        }
      },
      "/auth/login": {
        "post": {
          "tags": ["Auth"],
          "summary": "Login user",
          "description": "This route logs in a user by using an email and password.",
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "description": "Credentials",
              "required": true,
              "schema": {
                "$ref": "#/definitions/UserLogin"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Login successful",
              "schema": {
                "$ref": "#/definitions/LoginResponse"
              }
            },
            "400": {
              "description": "Missing email or password"
            },
            "404": {
              "description": "User not found"
            }
          }
        }
      },
...
    },
    "securityDefinitions": {
      "Bearer": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization"
      }
    },
    "definitions": {
      "UserRegister": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "UserLogin": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "user": {
            "type": "string"
          },
          "userID": {
            "type": "string"
          },
          "token": {
            "type": "string"
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "userID": {
            "type": "string"
          },
          "token": {
            "type": "string"
          }
        }
      },
...
    }
  }